Moved more code to parsing analysis
