+++ external-crates/move/crates/move-analyzer/src/analysis/parsing_analysis.rs
@@ -8,7 +8,7 @@ use crate::{
         cursor::{CursorContext, CursorDefinition, CursorPosition},
         ignored_function,
         mod_defs::{AutoImportInsertionInfo, AutoImportInsertionKind, CallInfo, ModuleDefs},
-        parsed_address, parsing_leading_and_mod_names_to_map_key, parsing_mod_def_to_map_key,
+        parsed_address,
         use_def::{References, UseDef, UseDefMap},
     },
     utils::{loc_end_to_lsp_position_opt, loc_start_to_lsp_position_opt},
@@ -922,6 +922,34 @@ impl<'a> ParsingAnalysisContext<'a> {
         // If the cursor is in this item, mark that down.
         update_cursor!(IDENT, self.cursor, field, FieldDefn);
     }
+}
+
+/// Produces module ident string of the form pkg::module to be used as a map key.
+/// It's important that these are consistent between parsing AST and typed AST.
+pub fn parsing_mod_def_to_map_key(
+    pkg_addresses: &NamedAddressMap,
+    mod_def: &P::ModuleDefinition,
+) -> Option<String> {
+    // we assume that modules are declared using the PkgName::ModName pattern (which seems to be the
+    // standard practice) and while Move allows other ways of defining modules (i.e., with address
+    // preceding a sequence of modules), this method is now deprecated.
+    //
+    // TODO: make this function simply return String when the other way of defining modules is
+    // removed
+    mod_def
+        .address
+        .map(|a| parsing_leading_and_mod_names_to_map_key(pkg_addresses, a, mod_def.name))
+}
+
+/// Produces module ident string of the form pkg::module to be used as a map key.
+/// It's important that these are consistent between parsing AST and typed AST.
+fn parsing_leading_and_mod_names_to_map_key(
+    pkg_addresses: &NamedAddressMap,
+    ln: P::LeadingNameAccess,
+    name: P::ModuleName,
+) -> String {
+    let parsed_addr = parsed_address(ln, pkg_addresses);
+    format!("{}::{}", parsed_addr, name).to_string()
 }
 
 /// Produces module ident string of the form pkg::module to be used as a map key.
